Finished achievement implementation
-Added "Daring Escape", "Save The World", "Absolute Victory", and "Swift Rescue" achievements

NOTE: due to additions to persistent data, there is a small chance that this might break compatibility with older save files. Best-case scenario is that older saves either work fine or are only locked out of the "Absolute Victory" and "Swift Rescue" achievements; worst-case is that older saves might be  completely incompatible.